+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Scoping dependencies to remove conflicts.
+### Added
+- Refund option to woocommerce back-office
+
 
 ## [1.1.2] - 2024-02-19